getmerchantavatar function is integrated to reimbursementitem.js
