+++ src/components/ReimbursementItem.js
@@ -1,12 +1,26 @@
 import React from "react";
 import { View, Text } from "react-native";
 import tw from "tailwind-react-native-classnames";
-import { getMerchantAvatar } from "../utils/Avatar";
+import { getRandomColor } from "../utils/common";
 
 const STATUS_COLORS = {
   Draft: "text-gray-500",
   "Approval Pending": "text-yellow-500",
   Approved: "text-green-500",
+};
+
+const getInitials = (name) => {
+  if (!name) return "";
+  const words = name.split(" ");
+  if (words.length === 1) return words[0].slice(0, 2).toUpperCase();
+  return (words[0][0] + words[1][0]).toUpperCase();
+};
+
+const getMerchantAvatar = (merchant) => {
+  return {
+    initials: getInitials(merchant),
+    backgroundColor: getRandomColor(),
+  };
 };
 
 const ReimbursementItem = ({ item }) => {
