+++ src/utils/common.js
@@ -1,0 +1,5 @@
+import { COLORS } from "../constants/colors";
+
+const getRandomColor = () => COLORS[Math.floor(Math.random() * COLORS.length)];
+
+export { getRandomColor };
